decouple default component from react root
